route-planner-backend/requirements.txt

aiohappyeyeballs==2.4.4 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.11.11 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.2 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.7.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.3.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
dataclasses-json==0.6.7 ; python_version >= "3.12" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi-cli[standard]==0.0.7 ; python_version >= "3.12" and python_version < "4.0"
fastapi[standard]==0.115.6 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
geographiclib==2.0 ; python_version >= "3.12" and python_version < "4.0"
geopy==2.4.1 ; python_version >= "3.12" and python_version < "4.0"
googlemaps==4.10.0 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.12" and python_version < "4.0"
httpx-sse==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
jiter==0.8.2 ; python_version >= "3.12" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.12" and python_version < "4.0"
jsonpointer==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
langchain-community==0.3.13 ; python_version >= "3.12" and python_version < "4.0"
langchain-core==0.3.28 ; python_version >= "3.12" and python_version < "4.0"
langchain-text-splitters==0.3.4 ; python_version >= "3.12" and python_version < "4.0"
langchain==0.3.13 ; python_version >= "3.12" and python_version < "4.0"
langgraph-checkpoint==2.0.9 ; python_version >= "3.12" and python_version < "4.0"
langgraph-sdk==0.1.48 ; python_version >= "3.12" and python_version < "4.0"
langgraph==0.2.60 ; python_version >= "3.12" and python_version < "4.0"
langsmith==0.2.7 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
marshmallow==3.23.2 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
msgpack==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.1.0 ; python_version >= "3.12" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
openai==1.58.1 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.13 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
propcache==0.2.1 ; python_version >= "3.12" and python_version < "4.0"
psycopg-binary==3.2.3 ; implementation_name != "pypy" and python_version >= "3.12" and python_version < "4.0"
psycopg[binary]==3.2.3 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.4 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rich-toolkit==0.12.0 ; python_version >= "3.12" and python_version < "4.0"
rich==13.9.4 ; python_version >= "3.12" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.36 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.41.3 ; python_version >= "3.12" and python_version < "4.0"
tenacity==9.0.0 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.12" and python_version < "4.0"
typer==0.15.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.21.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "4.0"
watchfiles==1.0.3 ; python_version >= "3.12" and python_version < "4.0"
websockets==14.1 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.18.3 ; python_version >= "3.12" and python_version < "4.0"
